% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportar_dados.R
\name{exportar_csv}
\alias{exportar_csv}
\title{Exportar bases em formato .csv separado por ';' e com enconding UTF-8.}
\usage{
exportar_csv(caminho_para_salvar_bases, ...)
}
\arguments{
\item{caminho_para_salvar_bases}{caminho da pasta para salvar o arquivo, como
"C:/documentos/" ou "./" para savar no projeto ativo}

\item{...}{parametros adicionais para a exportacao do csv}
}
\value{
invisivel
}
\description{
Exportar bases em formato .csv separado por ';' e com enconding UTF-8.
}
