% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportar_dados.R
\name{exportar_xlsx}
\alias{exportar_xlsx}
\title{Exportar bases em formato .xlsx (Excel Files).}
\usage{
exportar_xlsx(caminho_para_salvar_bases, ...)
}
\arguments{
\item{caminho_para_salvar_bases}{caminho da pasta para salvar o arquivo, como
"C:/documentos/" ou "./" para savar no projeto ativo}

\item{...}{parametros adicionais para a exportacao}
}
\value{
invisivel
}
\description{
Exporta as duas bases em planilhas distintas de uma pasta de trabalho excel.
}
