% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorar_mapeamentos.R
\name{generos_plantios_disponiveis}
\alias{generos_plantios_disponiveis}
\title{Generos existentes nos mapeamentos disponiveis}
\usage{
generos_plantios_disponiveis(exibir_nome_mapeamento = FALSE)
}
\arguments{
\item{exibir_nome_mapeamento}{padrao = FALSE, se TRUE exibe o nome do mapeamento}
}
\value{
Uma tibble
}
\description{
Resume quais sao os generos dos plantios florestais existentes nas bases utilizadas
no pacote.
}
