% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorar_mapeamentos.R
\name{mapeamento_existente_uf}
\alias{mapeamento_existente_uf}
\title{Mapeamentos/Relatorios disponiveis para determinada unidade federativa (estado)}
\usage{
mapeamento_existente_uf(unidade_federativa = "PR")
}
\arguments{
\item{unidade_federativa}{Sigla da Unidade federativa do Brasil, se nao for
definida nenhuma, a funcao ira retornar os mapeamentos para o Parana - PR.}
}
\value{
Retorna uma tible organizada por ano, do mais recente para o mais antigo,
com areas totais por mapeamento, informando qual a abordagem do mesmo.
}
\description{
Mapeamentos/Relatorios disponiveis para determinada unidade federativa (estado)
}
