% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorar_mapeamentos.R
\name{mapeamentos_disponiveis}
\alias{mapeamentos_disponiveis}
\title{Mapeamentos/Relatorios disponiveis dentro do pacote}
\usage{
mapeamentos_disponiveis()
}
\value{
Uma tibble com 3 colunas
}
\description{
Usando as duas bases existentes no pacote, resume quais sao os mapeamentos e
relatorios disponiveis, alem de informar qual a fonte dos mesmos.
}
