% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapeamentos_municipios}
\alias{mapeamentos_municipios}
\title{Dados sobre os mapeamentos em nível municipal de plantios florestais no Brasil}
\format{
Uma tibble com 32,588 linhas e 12 colunas, contendo:
\describe{
\item{mapeamento}{Nome do mapeamento ou relatório de origem}
\item{fonte}{Fonte dos dados, não é necessariamente o responsável pelo
relatório}
\item{ano_base}{Ano base do mapeamento, diferere do ano que o relatório foi
publicado, geralmente}
\item{uf}{Unidade Federativa da União}
\item{estado}{Nome do Estado da União}
\item{municipio}{Nome do município}
\item{code_muni}{Código único do município pelo IBGE}
\item{genero}{Gênero do plantio florestal}
\item{area_ha}{Área plantada (ha)}
}
}
\source{
Diversos - \url{https://github.com/maykongpedro/2021-07-04-extracao-mapeamentos-plantios-florestais}
}
\usage{
mapeamentos_municipios
}
\description{
Base de dados contendo os mapeamentos de plantios florestais no Brasil,
a nível de município.
}
\examples{
head(mapeamentos_municipios)
}
\keyword{datasets}
