% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos.R
\name{plotar_historico_brasil}
\alias{plotar_historico_brasil}
\title{Title}
\usage{
plotar_historico_brasil(mapeamento = "IBA", genero = "Todos")
}
\arguments{
\item{genero}{}
}
\value{

}
\description{
Title
}
