% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos.R
\name{plotar_historico_iba}
\alias{plotar_historico_iba}
\title{Plotar Grafico de Historico com os dados do IBA}
\usage{
plotar_historico_iba(abrangecia_uf = NULL, exibir_rotulos = FALSE)
}
\arguments{
\item{abrangecia_uf}{Filtro de estado, como "PR", por exemplo. O padrao e NULL,
para trazer dados do Brasil inteiro.}

\item{exibir_rotulos}{TRUE ou FALSE. Se TRUE exibe os rotulos em cima das colunas.}
}
\value{
Um grafico de colunas por ano, da area em milhoes de hectares.
}
\description{
Realiza um plot simples com os dados de mapeamento do IBA, tanto dos dados
disponibilizados pelo SNIF quanto dos dados apresentados no Relatorio Anual 2020.
O plot pode ser para um estado especifico ou geral.
}
\details{
A funcao pode demorar um pouco para demonstrar o resultado.
}
