% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorar_mapeamentos.R
\name{serie_historicas_disponiveis}
\alias{serie_historicas_disponiveis}
\title{Area dos Mapeamentos/Relatorios com series historicas por ano}
\usage{
serie_historicas_disponiveis()
}
\value{
Uma tibble com os mapeamentos que possuem uma serie historica
}
\description{
Area dos Mapeamentos/Relatorios com series historicas por ano
}
