% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorar_mapeamentos.R
\name{series_historicas_disponiveis}
\alias{series_historicas_disponiveis}
\title{Area dos Mapeamentos/Relatorios com series historicas por ano}
\usage{
series_historicas_disponiveis()
}
\value{
Uma tibble com os mapeamentos que possuem uma serie historica,
com a respectiva area para cada ano-base
}
\description{
Area dos Mapeamentos/Relatorios com series historicas por ano
}
